+++ src/components/Goals.tsx
@@ -1,18 +1,16 @@
-/* eslint-disable @typescript-eslint/no-unused-vars */
 import React, { useEffect, useMemo, useState } from 'react';
 import {
   Avatar, Box, Button, Card, Chip,
-  Input, Sheet, Stack, Textarea, Typography,
+  Sheet, Stack, Typography,
 } from '@mui/joy';
-import robot from 'assets/robot_taking_notes.webp';
 import { TbHeartHandshake } from 'react-icons/tb';
 import { ColorPaletteProp, SxProps } from '@mui/joy/styles/types';
-import { MdOutlineRocketLaunch, MdSend } from 'react-icons/md';
+import { MdOutlineRocketLaunch } from 'react-icons/md';
 import { HiOutlineSparkles } from 'react-icons/hi2';
 import { Parallax } from 'react-scroll-parallax';
 import { animated, useSpringValue } from '@react-spring/web';
-import { FaFire } from 'react-icons/fa';
 import { Default, Mobile, useMobileMode } from './Responsive';
+import Reach from './Reach';
 
 function Grid({ sx }: {sx?: SxProps}) {
   return (
@@ -115,8 +113,8 @@ function Cursor({ step } : {step: number}) {
     rotate: '-90deg',
     opacity: '1',
   }, {
-    top: '76%',
-    left: '15%',
+    top: '75%',
+    left: '4%',
     rotate: '0deg',
     opacity: '1',
   }, {
@@ -403,122 +401,6 @@ function Board({ step } : {step: number}) {
         <Comment step={step} />
       </Card>
     </Sheet>
-  );
-}
-
-function Reach({ step } : {step: number}) {
-  const mobile = useMobileMode();
-
-  const opacity = useSpringValue(0);
-  const top = useSpringValue('80%');
-
-  useEffect(() => {
-    opacity.start(step >= 4 ? 1 : 0);
-    top.start(step >= 4 ? '67%' : '80%');
-  }, [step]);
-
-  return (
-    <Box
-      component={animated.form}
-      action="https://api.web3forms.com/submit"
-      method="POST"
-      onSubmit={() => {
-        opacity.start(0);
-        top.start('80%');
-      }}
-      sx={{
-        position: 'absolute',
-        left: 0,
-        display: 'flex',
-        flexDirection: 'row',
-        justifyContent: mobile ? 'center' : 'flex-start',
-        paddingX: mobile ? '0' : '5%',
-        alignItems: 'center',
-        width: mobile ? 'calc(100% + 2rem)' : '100%',
-        maxHeight: '27rem',
-        marginX: mobile ? '-1rem' : 'auto',
-      }}
-      style={{
-        opacity,
-        top,
-      }}
-    >
-
-      <Stack sx={{
-        width: 'min(30rem, 90%)',
-        display: 'flex',
-        flexDirection: 'column',
-        gap: 1,
-      }}
-      >
-        <input type="hidden" name="access_key" value="4e4e6aee-a458-4774-a6e3-a6df6c19abe5" />
-        <Typography level="h1">
-          Let&apos;s
-          {' '}
-          <Typography color="danger">
-            connect.
-          </Typography>
-        </Typography>
-        <Input
-          variant="plain"
-          placeholder="john@acme.co.uk"
-          type="email"
-          name="email"
-          required
-        />
-        <Input
-          variant="plain"
-          placeholder="John Doe"
-          type="text"
-          name="name"
-          required
-        />
-        <Textarea
-          variant="plain"
-          placeholder="Your message here..."
-          name="message"
-          required
-          minRows={4}
-        />
-        <Stack direction="row" justifyContent="end" gap={1}>
-          <Button
-            component="a"
-            color="danger"
-            variant="soft"
-            href="https://www.youtube.com/watch?v=dQw4w9WgXcQ"
-            target="_blank"
-            startDecorator={(
-              <FaFire />
-        )}
-          >
-            Surprise me
-          </Button>
-          <Button
-            type="submit"
-            variant="soft"
-            startDecorator={(
-              <MdSend />
-          )}
-          >
-            Submit
-          </Button>
-        </Stack>
-      </Stack>
-      <Default>
-        <img
-          src={robot}
-          alt="robot taking notes"
-          style={{
-            position: 'absolute',
-            top: '-5rem',
-            left: 'max(65%, 32rem)',
-            height: '30rem',
-            opacity: 0.8,
-            filter: 'drop-shadow(0 -1rem 20px #feffee37) drop-shadow(2rem 3rem 40px #5074a33f) contrast(1.2)',
-          }}
-        />
-      </Default>
-    </Box>
   );
 }
